packages/core/pyproject.toml
[tool.poetry]
name = "core"
version = "1.0.0"
description = "The core module of this project contains the main business logic of the application and can be used across different apps"
authors = ["Ken Yip <ken20206@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8.17"
sqlalchemy = "^2.0.16"
alembic = "^1.11.1"
dynaconf = "^3.1.12"
python-dotenv = "^1.0.0"
uuid = "^1.30"
pymysql = "^1.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"
faker = "^18.10.1"
autoflake = "^2.1.1"
isort = "^5.12.0"
flake8 = "^6.0.0"
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
